enforcer(): Use abort() instead of exit(255)
By using abort(), the program has a chance to produce a core dump,
which one can use to determine where from the hard-coded /tmp stuff
was called from. It also makes it possible to determine which program
had this bug, by examining the core, a simple exit in a longer
pipeline may not be as helpful.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>