Build system simplification
tmperamental's build system is a bit of a kludge, in that it is
needlessly complex, and also does a few things wrong (setting the
soversion and the symlinks on an LD_PRELOAD-only thing).

This patch completely redoes the build system, replacing it with a
single top-level Makefile that will build everything into an 'out/'
directory. It also stops setting the soname, nor does it create
symlinks. It does make it possible to override the pkglibdir, so that
the package is easily multi-archifiable.

Oh, and include/tmperamental.h is gone, it wasn't useful either.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>